pyproject.toml
[project]
name = "agentic-knowledge-system"
version = "0.1.0"
description = "An agentic system for knowledge retrieval and query processing using AutoGen"
requires-python = ">=3.9"
dependencies = [
    "pyautogen>=0.2.0",
    "anthropic>=0.8.0",
    "google-generativeai>=0.3.0",
    "boto3>=1.28.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",
    "pyyaml>=6.0.1",
    "sqlalchemy>=2.0.0",
    "databricks-sql-connector>=2.9.0",
    "gql>=3.4.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "redis>=5.0.0",
    "tenacity>=8.2.0",
    "loguru>=0.7.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"] 